moved cmd line parsing  into separate source file
